split application class into cpp/hpp pair
